backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-multipart==0.0.6
pillow==10.1.0
pytesseract==0.3.10
instaloader==4.10.1
httpx==0.25.2
beautifulsoup4==4.12.2
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
alembic==1.13.0
aiofiles==23.2.1
